tools/fastapi/auth-api/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = ""
authors = ["Yann Beauxis <pro@yannbeauxis.net>"]
readme = "README.md"
packages = [{include = "auth_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.1"}
python-jose = "^3.3.0"
magic-admin = "^0.3.3"

[tool.poetry.group.test.dependencies]
black = "^23.3.0"
ruff = "^0.0.263"
mypy = "^1.2.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
types-python-jose = "^3.3.4.7"

[tool.poetry.group.dev.dependencies]
rope = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
